Put domain for todo row in a different file.
